requirements.txt
numpy<2
openai-whisper
pandas==2.2.2
openpyxl==3.1.5
jiwer==3.1.0
tqdm==4.67.1
